+++ src/core/Table/Table.tsx
@@ -19,11 +19,11 @@ import {
   ActionType,
   TableInstance,
   useExpanded,
+  usePagination,
 } from 'react-table';
 import { ProgressRadial } from '../ProgressIndicators';
-import { useTheme } from '../utils/hooks/useTheme';
+import { useTheme, CommonProps } from '../utils';
 import '@itwin/itwinui-css/css/table.css';
-import { CommonProps } from '../utils/props';
 import SvgSortDown from '@itwin/itwinui-icons-react/cjs/icons/SortDown';
 import SvgSortUp from '@itwin/itwinui-icons-react/cjs/icons/SortUp';
 import { getCellStyle } from './utils';
@@ -44,6 +44,39 @@ import {
 import { onSingleSelectHandler } from './actionHandlers/selectHandler';
 
 const singleRowSelectedAction = 'singleRowSelected';
+
+export type TablePaginatorRendererProps = {
+  /**
+   * The zero-based index of the current page.
+   */
+  currentPage: number;
+  /**
+   * Total number of rows.
+   */
+  totalRowsCount: number;
+  /**
+   * Number of rows per page.
+   */
+  pageSize: number;
+  /**
+   * Callback when page is changed.
+   */
+  onPageChange: (page: number) => void;
+  /**
+   * Callback when page size is changed.
+   */
+  onPageSizeChange: (size: number) => void;
+  /**
+   * Modify the size of the pagination (adjusts the elements size).
+   * @default 'default' if Table density is `default` else `small`
+   */
+  size?: 'default' | 'small';
+  /**
+   * Flag whether data is still loading and total rows count is not known.
+   * @default false
+   */
+  isLoading?: boolean;
+};
 
 /**
  * Table props.
@@ -145,6 +178,26 @@ export type TableProps<
    * @default 'default'
    */
   density?: 'default' | 'condensed' | 'extra-condensed';
+  /**
+   * Flag whether to select a row when clicked anywhere inside of it.
+   * @default true
+   */
+  selectRowOnClick?: boolean;
+  /**
+   * Function that takes `TablePaginatorRendererProps` as an argument and returns pagination component.
+   *
+   * Recommended to use `TablePaginator`. Passing `props` to `TablePaginator` handles all state management and is enough for basic use-cases.
+   * @example
+   * (props: TablePaginatorRendererProps) => (
+   *   <TablePaginator {...props} />
+   * )
+   */
+  paginatorRenderer?: (props: TablePaginatorRendererProps) => React.ReactNode;
+  /**
+   * Number of rows per page.
+   * @default 25
+   */
+  pageSize?: number;
 } & Omit<CommonProps, 'title'>;
 
 /**
@@ -222,6 +275,9 @@ export const Table = <
     density = 'default',
     selectSubRows = true,
     getSubRows,
+    selectRowOnClick = true,
+    paginatorRenderer,
+    pageSize = 25,
     ...rest
   } = props;
 
@@ -303,6 +359,8 @@ export const Table = <
 
   const instance = useTable<T>(
     {
+      manualPagination: !paginatorRenderer, // Prevents from paginating rows in regular table without pagination
+      paginateExpandedRows: false, // When false, it shows sub-rows in the current page instead of splitting them
       ...props,
       columns,
       defaultColumn,
@@ -312,12 +370,14 @@ export const Table = <
       selectSubRows,
       data,
       getSubRows,
+      initialState: { pageSize, ...props.initialState },
     },
     useFlexLayout,
     useFilters,
     useSubRowFiltering(hasAnySubRows),
     useSortBy,
     useExpanded,
+    usePagination,
     useRowSelect,
     useSubRowSelection,
     useExpanderCell(subComponent, expanderCell, isRowDisabled),
@@ -334,6 +394,9 @@ export const Table = <
     allColumns,
     filteredFlatRows,
     dispatch,
+    page,
+    gotoPage,
+    setPageSize,
   } = instance;
 
   const ariaDataAttributes = Object.entries(rest).reduce(
@@ -351,7 +414,7 @@ export const Table = <
   const onRowClickHandler = React.useCallback(
     (event: React.MouseEvent, row: Row<T>) => {
       const isDisabled = isRowDisabled?.(row.original);
-      if (isSelectable && !isDisabled) {
+      if (isSelectable && !isDisabled && selectRowOnClick) {
         if (!row.isSelected && !event.ctrlKey) {
           dispatch({
             type: singleRowSelectedAction,
@@ -361,123 +424,159 @@ export const Table = <
           row.toggleRowSelected(!row.isSelected);
         }
       }
-      !isDisabled && onRowClick?.(event, row);
+      if (!isDisabled) {
+        onRowClick?.(event, row);
+      }
     },
-    [dispatch, isSelectable, onRowClick, isRowDisabled],
+    [isRowDisabled, isSelectable, selectRowOnClick, dispatch, onRowClick],
+  );
+
+  React.useEffect(() => {
+    setPageSize(pageSize);
+  }, [pageSize, setPageSize]);
+
+  const paginatorRendererProps: TablePaginatorRendererProps = React.useMemo(
+    () => ({
+      currentPage: state.pageIndex,
+      pageSize: state.pageSize,
+      totalRowsCount: rows.length,
+      size: density !== 'default' ? 'small' : 'default',
+      isLoading,
+      onPageChange: gotoPage,
+      onPageSizeChange: setPageSize,
+    }),
+    [
+      density,
+      gotoPage,
+      isLoading,
+      rows.length,
+      setPageSize,
+      state.pageIndex,
+      state.pageSize,
+    ],
   );
 
   return (
-    <div
-      ref={(element) => setOwnerDocument(element?.ownerDocument)}
-      id={id}
-      {...getTableProps({
-        className: cx(
-          'iui-table',
-          { [`iui-${density}`]: density !== 'default' },
-          className,
-        ),
-        style,
-      })}
-      {...ariaDataAttributes}
-    >
-      <div className='iui-table-header'>
-        {headerGroups.slice(1).map((headerGroup: HeaderGroup<T>) => {
-          const headerGroupProps = headerGroup.getHeaderGroupProps({
-            className: 'iui-row',
-          });
-          return (
-            <div {...headerGroupProps} key={headerGroupProps.key}>
-              {headerGroup.headers.map((column) => {
-                const columnProps = column.getHeaderProps({
-                  ...column.getSortByToggleProps(),
-                  className: cx(
-                    'iui-cell',
-                    { 'iui-actionable': column.canSort },
-                    { 'iui-sorted': column.isSorted },
-                    column.columnClassName,
-                  ),
-                  style: { ...getCellStyle(column) },
-                });
-                return (
-                  <div {...columnProps} key={columnProps.key} title={undefined}>
-                    {column.render('Header')}
-                    {!isLoading && (data.length != 0 || areFiltersSet) && (
-                      <FilterToggle
-                        column={column}
-                        ownerDocument={ownerDocument}
-                      />
-                    )}
-                    {!isLoading && data.length != 0 && column.canSort && (
-                      <div className='iui-cell-end-icon'>
-                        {column.isSorted && column.isSortedDesc ? (
-                          <SvgSortUp
-                            className='iui-icon iui-sort'
-                            aria-hidden
-                          />
-                        ) : (
-                          <SvgSortDown
-                            className='iui-icon iui-sort'
-                            aria-hidden
-                          />
-                        )}
-                      </div>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-          );
+    <>
+      <div
+        ref={(element) => setOwnerDocument(element?.ownerDocument)}
+        id={id}
+        {...getTableProps({
+          className: cx(
+            'iui-table',
+            { [`iui-${density}`]: density !== 'default' },
+            className,
+          ),
+          style,
         })}
-      </div>
-      <div {...getTableBodyProps({ className: 'iui-table-body' })}>
-        {data.length !== 0 &&
-          rows.map((row: Row<T>) => {
-            prepareRow(row);
+        {...ariaDataAttributes}
+      >
+        <div className='iui-table-header'>
+          {headerGroups.slice(1).map((headerGroup: HeaderGroup<T>) => {
+            const headerGroupProps = headerGroup.getHeaderGroupProps({
+              className: 'iui-row',
+            });
             return (
-              <TableRowMemoized
-                row={row}
-                rowProps={rowProps}
-                isLast={row.index === data.length - 1}
-                onRowInViewport={onRowInViewportRef}
-                onBottomReached={onBottomReachedRef}
-                intersectionMargin={intersectionMargin}
-                state={state}
-                key={row.getRowProps().key}
-                onClick={onRowClickHandler}
-                subComponent={subComponent}
-                isDisabled={!!isRowDisabled?.(row.original)}
-                tableHasSubRows={hasAnySubRows}
-                tableInstance={instance}
-                expanderCell={expanderCell}
-              />
+              <div {...headerGroupProps} key={headerGroupProps.key}>
+                {headerGroup.headers.map((column) => {
+                  const columnProps = column.getHeaderProps({
+                    ...column.getSortByToggleProps(),
+                    className: cx(
+                      'iui-cell',
+                      { 'iui-actionable': column.canSort },
+                      { 'iui-sorted': column.isSorted },
+                      column.columnClassName,
+                    ),
+                    style: { ...getCellStyle(column) },
+                  });
+                  return (
+                    <div
+                      {...columnProps}
+                      key={columnProps.key}
+                      title={undefined}
+                    >
+                      {column.render('Header')}
+                      {!isLoading && (data.length != 0 || areFiltersSet) && (
+                        <FilterToggle
+                          column={column}
+                          ownerDocument={ownerDocument}
+                        />
+                      )}
+                      {!isLoading && data.length != 0 && column.canSort && (
+                        <div className='iui-cell-end-icon'>
+                          {column.isSorted && column.isSortedDesc ? (
+                            <SvgSortUp
+                              className='iui-icon iui-sort'
+                              aria-hidden
+                            />
+                          ) : (
+                            <SvgSortDown
+                              className='iui-icon iui-sort'
+                              aria-hidden
+                            />
+                          )}
+                        </div>
+                      )}
+                    </div>
+                  );
+                })}
+              </div>
             );
           })}
-        {isLoading && data.length === 0 && (
-          <div className={'iui-table-empty'}>
-            <ProgressRadial indeterminate={true} />
-          </div>
-        )}
-        {isLoading && data.length !== 0 && (
-          <div className='iui-row'>
-            <div className='iui-cell' style={{ justifyContent: 'center' }}>
-              <ProgressRadial
-                indeterminate={true}
-                size='small'
-                style={{ float: 'none', marginLeft: 0 }}
-              />
+        </div>
+        <div {...getTableBodyProps({ className: 'iui-table-body' })}>
+          {data.length !== 0 &&
+            page.map((row: Row<T>) => {
+              prepareRow(row);
+              return (
+                <TableRowMemoized
+                  row={row}
+                  rowProps={rowProps}
+                  isLast={row.index === data.length - 1}
+                  onRowInViewport={onRowInViewportRef}
+                  onBottomReached={onBottomReachedRef}
+                  intersectionMargin={intersectionMargin}
+                  state={state}
+                  key={row.getRowProps().key}
+                  onClick={onRowClickHandler}
+                  subComponent={subComponent}
+                  isDisabled={!!isRowDisabled?.(row.original)}
+                  tableHasSubRows={hasAnySubRows}
+                  tableInstance={instance}
+                  expanderCell={expanderCell}
+                />
+              );
+            })}
+          {isLoading && data.length === 0 && (
+            <div className={'iui-table-empty'}>
+              <ProgressRadial indeterminate={true} />
             </div>
-          </div>
-        )}
-        {!isLoading && data.length === 0 && !areFiltersSet && (
-          <div className={'iui-table-empty'}>{emptyTableContent}</div>
-        )}
-        {!isLoading &&
-          (data.length === 0 || filteredFlatRows.length === 0) &&
-          areFiltersSet && (
-            <div className={'iui-table-empty'}>{emptyFilteredTableContent}</div>
           )}
+          {isLoading && data.length !== 0 && (
+            <div className='iui-row'>
+              <div className='iui-cell' style={{ justifyContent: 'center' }}>
+                <ProgressRadial
+                  indeterminate={true}
+                  size='small'
+                  style={{ float: 'none', marginLeft: 0 }}
+                />
+              </div>
+            </div>
+          )}
+          {!isLoading && data.length === 0 && !areFiltersSet && (
+            <div className={'iui-table-empty'}>{emptyTableContent}</div>
+          )}
+          {!isLoading &&
+            (data.length === 0 || filteredFlatRows.length === 0) &&
+            areFiltersSet && (
+              <div className={'iui-table-empty'}>
+                {emptyFilteredTableContent}
+              </div>
+            )}
+        </div>
+        {paginatorRenderer?.(paginatorRendererProps)}
       </div>
-    </div>
+    </>
   );
 };